research_agent_system/requirements.txt
google-generativeai>=0.3.0
aiohttp>=3.8.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0
opentelemetry-instrumentation-aiohttp-client>=0.41b0
structlog>=23.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pydantic>=2.0.0
python-dotenv>=1.0.0
asyncio-mqtt>=0.1.0
pytest-cov>=4.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic-settings>=2.0.0
prometheus-client>=0.17.0
redis>=4.5.0
pymongo>=4.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
jinja2>=3.1.0